requirements.txt
Flask==1.1.4
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
itsdangerous==1.1.0
jinja2<3.1.0
markupsafe==1.1.1
psycopg2
python-dotenv
SQLAlchemy==1.3.13